refactor: merge IntentionIntrospector into IntentionAwarePolicyGraph
